% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/prediction.R
\name{postTraj}
\alias{postTraj}
\title{postTraj}
\usage{
postTraj(model, m, ids, ...)
}
\arguments{
\item{model}{fitted model object}

\item{m}{m for \code{posterior_traj} function}

\item{ids}{value of id}

\item{...}{other parameter option, see \code{posterior_traj}}
}
\value{
list of predicted values for the given id
}
\description{
prediction of the posterio trajectory for longitudinal marker while using rstanarm for Big data
}
\examples{
 \donttest{
##
library(survival)
library(dplyr)
fit6<-jmstanBig(dtlong=long2,dtsurv = surv2,longm=y~ x7+visit+(1|id),
         survm=Surv(time,status)~x1+visit,samplesize=200,time_var='visit',id='id')
P2<-postTraj(model<-fit6,m<-1,ids<-c(1,2,100))
pp1<-plot(P2$p1[[1]],plot_observed = TRUE)
pp2<-plot(P2$p1[[2]],plot_observed = TRUE)
pp3<-plot(P2$p1[[3]],plot_observed = TRUE)
##
   }
}
