% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/print.R
\name{print.jmcsBig}
\alias{print.jmcsBig}
\title{print.jmcsBig}
\usage{
\method{print}{jmcsBig}(x, ...)
}
\arguments{
\item{x}{fitted object}

\item{...}{others}
}
\value{
prints table containing various parameter estimates,
        SE, P- value for both survival and longitudinal submodel,
        if the model is bayesian it includes their credible interval too.
}
\description{
print method for class 'jmcsBig'
}
\examples{
 \donttest{
##
library(survival)
library(dplyr)
################################
mod2<-jmcsBig(dtlong=data.frame(long2),
dtsurv = data.frame(surv2),
longm=y~ x7+visit,
survm=Surv(time,status)~x1+visit,
rd= ~ visit|id,
samplesize=200,id='id')
print(mod2)
   }
}
