% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/print.R
\name{print.joinRMLBig}
\alias{print.joinRMLBig}
\title{print.joinRMLBig}
\usage{
\method{print}{joinRMLBig}(x, ...)
}
\arguments{
\item{x}{fitted object}

\item{...}{others}
}
\value{
prints table containing various parameter estimates,
        SE, P- value for both survival and longitudinal submodel,
        if the model is bayesian it includes their credible interval too.
}
\description{
print method for class 'joinRMLBig'
}
\examples{
 \donttest{
##
library(survival)
library(dplyr)
mod4<-joinRMLBig(dtlong=long2,
dtsurv = surv2,
longm=y~ x7+visit,
survm=Surv(time,status)~x1+visit,
rd=~ visit|id,
timeVar='visit',
samplesize=200,
id='id')
print(mod4)
   }
}
