% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/print.R
\name{print}
\alias{print}
\title{print}
\usage{
print(object, ...)
}
\arguments{
\item{object}{object}

\item{...}{others}
}
\value{
prints table containing various parameter estimates,
        SE, P- value for both survival and longitudinal submodel,
        if the model is bayesian it includes their credible interval too.
}
\description{
print
}
\examples{

 \donttest{
##
library(survival)
library(dplyr)
fit7<-jmstanBig(dtlong=long2,dtsurv = surv2,longm=y~ x7+visit+(1|id),
survm=Surv(time,status)~x1+visit,samplesize=200,time_var='visit',id='id')
################################
fit8<-jmcsBig(dtlong=data.frame(long2),dtsurv = data.frame(surv2),
longm=y~ x7+visit,survm=Surv(time,status)~x1+visit,rd= ~ visit|id,
samplesize=200,id='id')
#################################
fit9<-jmbayesBig(dtlong=long2,dtsurv = surv2 ,
longm=y~ x7+visit,survm=Surv(time,status)~x1+visit,
rd= ~ visit|id,timeVar='visit',nchain=1,samplesize=200,id='id')
#################################

fit10<-joinRMLBig(dtlong=long2,dtsurv = surv2,
longm=y~ x7+visit,survm=Surv(time,status)~x1+visit,
rd=~ visit|id,timeVar='visit',samplesize=200,id='id')
}
}
