% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dataset.R
\docType{data}
\name{surv2}
\alias{surv2}
\title{survival data}
\format{
a tibble of 13 columns and 1000 observations,
\describe{
\item{id}{id value for subjects}
\item{status}{survival status}
\item{time}{ survival time}
\item{visit}{visit time of longitudinal measurements}
\item{x1,x2,...,x7}{ different numeric and categorical variable}

}
}
\usage{
data(surv2)
}
\description{
A survival dataset related the long2 dataset, with different numeric and categorical covariate
}
\keyword{datasets}
