% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/print.R
\name{print.jmstanBig}
\alias{print.jmstanBig}
\title{print.jmstanBig}
\usage{
\method{print}{jmstanBig}(x, ...)
}
\arguments{
\item{x}{fitted object}

\item{...}{others}
}
\value{
prints table containing various parameter estimates,
        SE, P- value for both survival and longitudinal submodel,
        if the model is bayesian it includes their credible interval too.
}
\description{
print method for class 'jmstanBig'
}
\examples{

 \donttest{
##
library(survival)
library(dplyr)
mod1<-jmstanBig(dtlong=long2,
         dtsurv = surv2,
         longm=y~ x7+visit+(1|id),
         survm=Surv(time,status)~x1+visit,
         samplesize=200,
         time_var='visit',id='id')
print(mod1)
}
}
